refactor(calisma_karti): split prompts.ts into smaller modular files and fix single material UI
