+++ erpnextkta/public/view-calisma-karti/composables/prompts/durus.ts
@@ -1,0 +1,14 @@
+import { applyDecimalInputMode } from "./common";
+
+export function durusFields(reasonOptions: string) {
+    return applyDecimalInputMode([
+        {
+            fieldtype: "Select",
+            label: __("Duruş Nedeni"),
+            fieldname: "durus_nedeni",
+            reqd: 1,
+            options: reasonOptions
+        },
+        { fieldtype: "Small Text", label: __("Açıklama"), fieldname: "aciklama" }
+    ]);
+}
